+++ src/SubchannelSource.h
@@ -42,7 +42,7 @@ class SubchannelSource : public ModInput
 {
 public:
     SubchannelSource(
-            uint8_t sad,
+            uint16_t sad,
             uint16_t stl,
             uint8_t tpl
             );
